+++ Domain-Module/Auth-Module/Auth-Application/src/main/java/com/pawith/authapplication/service/command/feign/response/GoogleUserInfo.java
@@ -1,0 +1,16 @@
+package com.pawith.authapplication.service.command.feign.response;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class GoogleUserInfo {
+    private String email;
+    private String name;
+    public GoogleUserInfo(String email, String name) {
+        this.email = email;
+        this.name = name;
+    }
+}
