+++ Domain-Module/Auth-Module/Auth-Application/src/main/java/com/pawith/authapplication/service/command/handler/impl/GoogleOAuthHandler.java
@@ -1,21 +1,21 @@
 package com.pawith.authapplication.service.command.handler.impl;
 
-import com.pawith.authapplication.service.command.handler.AuthHandler;
 import com.pawith.authapplication.dto.OAuthRequest;
 import com.pawith.authapplication.dto.OAuthUserInfo;
+import com.pawith.authapplication.service.command.feign.GoogleOAuthFeignClient;
+import com.pawith.authapplication.service.command.feign.response.GoogleUserInfo;
+import com.pawith.authapplication.service.command.handler.AuthHandler;
 import com.pawith.commonmodule.enums.Provider;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import org.springframework.web.reactive.function.client.WebClient;
 
 @Component
+@RequiredArgsConstructor
 public class GoogleOAuthHandler implements AuthHandler {
     private static final Provider OAUTH_TYPE = Provider.GOOGLE;
-    private static final String GOOGLE_OAUTH_USER_INFO_URL = "https://www.googleapis.com/oauth2/v3/userinfo";
-    public static final String GOOGLE_AUTHORIZATION = "Authorization";
     public static final String GOOGLE_AUTHORIZATION_BEARER = "Bearer ";
+
+    private final GoogleOAuthFeignClient googleOAuthFeignClient;
 
     @Override
     public OAuthUserInfo handle(OAuthRequest authenticationInfo) {
@@ -30,23 +30,6 @@ public class GoogleOAuthHandler implements AuthHandler {
     }
 
     private GoogleUserInfo getGoogleUserInfo(final String accessToken){
-        return WebClient.create(GOOGLE_OAUTH_USER_INFO_URL)
-            .get()
-            .header(GOOGLE_AUTHORIZATION, GOOGLE_AUTHORIZATION_BEARER+accessToken)
-            .retrieve()
-            .bodyToMono(GoogleUserInfo.class)
-            .block();
-    }
-
-    @Getter
-    @NoArgsConstructor(access = AccessLevel.PROTECTED)
-    private static class GoogleUserInfo{
-        private String email;
-        private String name;
-
-        public GoogleUserInfo(String email, String name) {
-            this.email = email;
-            this.name = name;
-        }
+        return googleOAuthFeignClient.getGoogleUserInfo(GOOGLE_AUTHORIZATION_BEARER+accessToken);
     }
 }
